Fixed the whitelist state not being persisted.
As described in GlowstoneMC/Glowstone#491, reloading or restarting the
server can cause the whitelist state to be restored against the command
control. This commit combats that by ensuring that the state set by the
whitelist command is saved for the next potential read of the state.

A side effect of this change is that there is now a setter in the server
configuration to change values. This may be useful in the future for
similar issues.