Spawn preload tweaks and setup.sh can execute
Completely disables the spawn chunk anchor when persist spawn is disabled.
When persist spawn is enabled and a new spawn is set, the anchor will move to that new spawn location.
Spawn setting also now checks whether the new set spawn is the same as the old one, and if so, it prevents the spawn from being set again.
Persisting the spawn is now all located in setKeepspawnInMemory, intead of also being included in the constructor.
There is a new setting to populate chunks when they are anchored. This prevents the delay experienced when entering an anchored chunk (like spawn) for the first time, at the cost of increased server load, which may be unnecessary, which is why it is disabled by default. The setting is populateAnchoredChunks.