Change some default config values
Update help link
Glowstone -> Glowstone++

Populate spawn chunks by default. This increases server load time significantly the first time a world loads, and increases it very slightly every time after that. If you have no more new worlds to load, change this setting to false. 

populate-anchored-chunks reduces the time for "downloading terrain" when a player joins a never before loaded world, and also removes the server stutter and freezing players experience when a world is loaded. This is accomplished by populating it before players join the server, on server start. Not only is it beneficial to have this true by default, but it is default Minecraft behavior.