Add variable block sounds
This is a mix of https://github.com/GlowstoneMC/Glowstone/pull/242 and https://github.com/GlowstoneMC/Glowstone/pull/130. Thanks to @coasterman10 and @BeYkeRYkt for their work!

This commit adds block sounds with a pitch deviation of +/- 0.25. It changes the default value for the sound to stone, unless the block uses the default blocktype. In that case, it uses the wood sound. Some of the block sounds are not set correctly, especially 1.8 stuff.