Optimize water and config
Cache config values. May not work on reload but reload is buggy anyway.
This is especially helpful for the new classic water setting, which was using a lot of CPU for
checking if classic water was enabled.
Also, now water does not request a tick if it is at the lowest strength and can't go down, as there
is no point to ticking it.