Fixes #422 and other structure misc fixes
Rollback block changes if a structure is not to be generated. This fixes containers generating even if the structure is not generated.

Change boolean logic to consider a piece to be successfully generated when it generates successfully..

Use Objects functional method to check for non null values when wrapping pieces in the structure bounding box.

Remove logging noise when generating structures.