Make furnaces show as burning when cooking/burning (fixes GH-474)
This is not a perfect fix. Clients will see the block as 'BURNING_FURNACE' type, but plugins using Block#getType() will return Material.FURNACE. This fix only sends block-change to players.

Might need a better fix that will actually set the type to BURNING_FURNACE, but for the moment, changing the block type would destroy the BlockEntity.